Restructured Painting Logic
