--- rust/src/layout/tests.rs
@@ -1,22 +1,0 @@
-// use std::{fs::read_to_string, path::PathBuf, slice::from_raw_parts, str::from_utf8_unchecked};
-
-// use crate::{CharsMap, Dimensions, Layout};
-// use usfm::{BookContents, parse};
-
-// static MANIFEST_DIR: &str = env!("CARGO_MANIFEST_DIR");
-
-// #[test]
-// fn layout() {
-//     // let file = PathBuf::from(MANIFEST_DIR).join("usfm/02-GENengwebpb.usfm");
-//     // let input = read_to_string(file).unwrap();
-//     // let book = parse(&input);
-//     // let mut paragraphs = book
-//     //     .contents
-//     //     .iter()
-//     //     .filter_map(|c| match c {
-//     //         BookContents::Paragraph { contents, .. } => Some(contents),
-//     //         _ => None,
-//     //     })
-//     //     .take(1);
-//     // let paragraph = paragraphs.next().unwrap();
-// }
